Switch to Pypy (#24)
Pypy is faster than CPython and should reduce timeouts when parsing large fragments. Unfortunately, it does not have all features of the reference implementation and does not yet implement Python 3.8.

singledispatch appears to be broken and all uses had to be removed.

black and mypy do not work with Pypy and had to be removed.